Move registered transaction types to constants
