+++ lib/transaction/specialtransactions/constants.js
@@ -1,3 +1,16 @@
 module.exports = {
   PUBKEY_ID_SIZE: 20,
+  registeredTransactionTypes: {
+    TRANSACTION_NORMAL : 0,
+    TRANSACTION_PROVIDER_REGISTER : 1,
+    TRANSACTION_PROVIDER_UPDATE_SERVICE : 2,
+    TRANSACTION_PROVIDER_UPDATE_REGISTRAR : 3,
+    TRANSACTION_PROVIDER_UPDATE_REVOKE : 4,
+    TRANSACTION_COINBASE : 5,
+    TRANSACTION_SUBTX_REGISTER : 8,
+    TRANSACTION_SUBTX_TOPUP : 9,
+    TRANSACTION_SUBTX_RESETKEY : 10,
+    TRANSACTION_SUBTX_CLOSEACCOUNT : 11,
+    TRANSACTION_SUBTX_TRANSITION : 12,
+  }
 };
