+++ lib/transaction/specialtransactions/payload/payloadparser.js
@@ -1,5 +1,5 @@
-const RegisteredPayloadTypes = require('./registeredtypes');
-const SubTxRegisterPayload = require('./payload/subtxregisterpayload');
+const RegisteredPayloadTypes = require('../constants').registeredTransactionTypes;
+const SubTxRegisterPayload = require('./subtxregisterpayload');
 
 /**
  *
