+++ lib/transaction/specialtransactions/subtxregister.js
@@ -1,5 +1,5 @@
 var Transaction = require('../transaction');
-var TRANSACTION_SUBTX_REGISTER = require('./registeredtypes').TRANSACTION_SUBTX_REGISTER;
+var TRANSACTION_SUBTX_REGISTER = require('./constants').registeredTransactionTypes.TRANSACTION_SUBTX_REGISTER;
 var constants = require('./constants');
 var Preconditions = require('../../util/preconditions');
 var BufferUtil = require('../../util/buffer');
