+++ lib/transaction/transaction.js
@@ -25,7 +25,8 @@ var Output = require('./output');
 var Script = require('../script');
 var PrivateKey = require('../privatekey');
 var BN = require('../crypto/bn');
-var registeredTransactionTypes = require('./specialtransactions/registeredtypes');
+var registeredTransactionTypes = require('./specialtransactions/constants').registeredTransactionTypes;
+var payloadParser = require('./specialtransactions/payload/payloadparser');
 
 /**
  * Represents a transaction, a set of inputs and outputs to change ownership of tokens
