+++ lib/transaction/transactionfactory.js
@@ -1,6 +1,6 @@
 const Transaction = require('./transaction');
 const SpecialTransactions = require('./specialtransactions');
-const RegisteredTypes = require('./specialtransactions/registeredtypes');
+const RegisteredTypes = require('./specialtransactions/constants').registeredTransactionTypes;
 
 /**
  * Returns special transaction based on transaction payload
@@ -21,5 +21,8 @@ module.exports = {
   parseTransaction: function parseTranasction(serializedTransaction) {
     var transaction = new Transaction(serializedTransaction);
     return getSpecialTransaction(transaction);
+  },
+  createSpecialTransaction: function createSpecialTransaction() {
+
   }
 };
